requirements.txt
fastapi==0.95.2
starlette==0.27.0
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==1.10.13
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
